Scene 직렬화 추가
기존과 다르게 오브젝트 종류 별로 Serializer class를 만들어서 관리하기로 변경

gameobject와 component serializer class 추가

작동 확인함

GameObject
- Component
- Component
GameObject
- Component
예전 방식은 위와 같았으나 예전과 다르게 각각의 오브젝트는 하나의 ptree로 만듬 게임오브젝트가 소유한 컴포넌트의 uuid만 직렬화 하도록 변경 자식 컴포넌트는 다음 ptree에 저장

예시)
GameObject
-component [uuid, uuid]
Component
Component
GameObject
-component [uuid]
Component

+++ REngine/src/rengine/core/ComponentManager.cpp
@@ -31,7 +31,7 @@ namespace rengine
 		}
 		else
 		{
-			auto _components = make_pair(component->GetType(), Components(m_componentsList.size(),  component->GetType()));
+			auto _components = make_pair(component->GetType(), Components((uint32)m_componentsList.size(),  component->GetType()));
 			
 			_components.second.AddComponent(component);
